@@ -7,10 +7,10 @@ import com.tomgibara.geom.transform.Transform;
 //TODO use from not at in cons?
 public final class Rect implements Geometric {
 
-	public static final Rect UNIT_SQUARE = new Rect(0f, 0f, 1f, 1f);
-	public static final Rect BASIS_SQUARE = new Rect(-1f, -1f, 1f, 1f);
+	public static final Rect UNIT_SQUARE = new Rect(0.0, 0.0, 1.0, 1.0);
+	public static final Rect BASIS_SQUARE = new Rect(-1.0, -1.0, 1.0, 1.0);
 
-	public static Rect atOrigin(float x, float y) {
+	public static Rect atOrigin(double x, double y) {
 		return new Rect(
 			Math.min(x, 0),
 			Math.min(y, 0),
@@ -18,9 +18,9 @@ public final class Rect implements Geometric {
 			Math.max(y, 0));
 	}
 
-	public static Rect centerAtOrigin(float width, float height) {
-		float hw = width * 0.5f;
-		float hh = height * 0.5f;
+	public static Rect centerAtOrigin(double width, double height) {
+		double hw = width * 0.5;
+		double hh = height * 0.5;
 		return new Rect(-hw, -hh, hw, hh);
 	}
 
@@ -29,16 +29,16 @@ public final class Rect implements Geometric {
 		return atOrigin(pt.x, pt.y);
 	}
 
-	public static Rect atCenter(Point pt, float width, float height) {
+	public static Rect atCenter(Point pt, double width, double height) {
 		if (pt == null) throw new IllegalArgumentException("null pt");
-		if (width < 0f) throw new IllegalArgumentException("negative width");
-		if (height < 0f) throw new IllegalArgumentException("negative height");
-		float hw = width * 0.5f;
-		float hh = height * 0.5f;
+		if (width < 0.0) throw new IllegalArgumentException("negative width");
+		if (height < 0.0) throw new IllegalArgumentException("negative height");
+		double hw = width * 0.5;
+		double hh = height * 0.5;
 		return new Rect(pt.x - hw, pt.y - hh, pt.x + hw, pt.y + hh);
 	}
 
-	public static Rect atPoints(float x1, float y1, float x2, float y2) {
+	public static Rect atPoints(double x1, double y1, double x2, double y2) {
 		return new Rect(
 			Math.min(x1, x2),
 			Math.min(y1, y2),
@@ -55,6 +55,14 @@ public final class Rect implements Geometric {
 	public static Rect atPoint(Point pt) {
 		if (pt == null) throw new IllegalArgumentException("null pt");
 		return new Rect(pt.x, pt.y, pt.x, pt.y);
+	}
+
+	public static Rect atPointWithDimensions(double x, double y, double dx, double dy) {
+		return atPoints(x, y, x + dx, y + dy);
+	}
+
+	public static Rect atPointWithDimensions(Point pt, Vector dimensions) {
+		return atPointWithDimensions(pt.x, pt.y, dimensions.x, dimensions.y);
 	}
 
 	public static Rect unionRect(Rect r1, Rect r2) {
@@ -81,35 +89,35 @@ public final class Rect implements Geometric {
 	public static Rect intersectionRect(Rect r1, Rect r2) {
 		if (r1 == null) throw new IllegalArgumentException("null r1");
 		if (r2 == null) throw new IllegalArgumentException("null r2");
-		float minX = Math.max(r1.minX, r2.minX);
-		float minY = Math.max(r1.minY, r2.minY);
-		float maxX = Math.min(r1.maxX, r2.maxX);
-		float maxY = Math.min(r1.maxY, r2.maxY);
+		double minX = Math.max(r1.minX, r2.minX);
+		double minY = Math.max(r1.minY, r2.minY);
+		double maxX = Math.min(r1.maxX, r2.maxX);
+		double maxY = Math.min(r1.maxY, r2.maxY);
 		return minX < maxX && minY < maxY ? new Rect(minX, minY, maxX, maxY) : null;
 	}
 
-	public final float minX;
-	public final float minY;
-	public final float maxX;
-	public final float maxY;
+	public final double minX;
+	public final double minY;
+	public final double maxX;
+	public final double maxY;
 
-	Rect(float minX, float minY, float maxX, float maxY) {
+	Rect(double minX, double minY, double maxX, double maxY) {
 		this.minX = minX;
 		this.minY = minY;
 		this.maxX = maxX;
 		this.maxY = maxY;
 	}
 
-	public float getWidth() {
+	public double getWidth() {
 		return maxX - minX;
 	}
 
-	public float getHeight() {
+	public double getHeight() {
 		return maxY - minY;
 	}
 
 	public Point getCenter() {
-		return new Point((minX + maxX) * 0.5f, (minY + maxY) * 0.5f);
+		return new Point((minX + maxX) * 0.5, (minY + maxY) * 0.5);
 	}
 
 	// if the width or height are zero
@@ -117,28 +125,28 @@ public final class Rect implements Geometric {
 		return minX == maxX || minY == maxY;
 	}
 
-	public Rect translate(float x, float y) {
+	public Rect translate(double x, double y) {
 		return x == 0 && y == 0 ? this : new Rect(minX + x, minY + y, maxX + x, maxY + y);
 	}
 
 	public Rect translateToOrigin() {
-		return minX == 0f && minY == 0f ? this : new Rect(0f, 0f, getWidth(), getHeight());
+		return minX == 0.0 && minY == 0.0 ? this : new Rect(0.0, 0.0, getWidth(), getHeight());
 	}
 
 	//TODO rename to outset and supply an inset too?
-	public Rect offset(float offset) {
-		if (offset == 0f) return this;
-		if (offset > 0f) return new Rect(minX - offset, minY - offset, maxX + offset, maxY + offset);
+	public Rect offset(double offset) {
+		if (offset == 0.0) return this;
+		if (offset > 0.0) return new Rect(minX - offset, minY - offset, maxX + offset, maxY + offset);
 		return Rect.atPoints(minX - offset, minY - offset, maxX + offset, maxY + offset);
 	}
 
 	//TODO put this on offset instead?
 	public Rect offset(Offset offset) {
 		if (offset == Offset.IDENTITY) return this;
-		float x1 = minX + offset.toMinX;
-		float y1 = minY + offset.toMinY;
-		float x2 = maxX + offset.toMaxX;
-		float y2 = maxX + offset.toMaxY;
+		double x1 = minX + offset.toMinX;
+		double y1 = minY + offset.toMinY;
+		double x2 = maxX + offset.toMaxX;
+		double y2 = maxX + offset.toMaxY;
 		//TODO do this more efficiently: if it flips hz or vt then slower path
 		// ie if hz. growth + width < 0 flip x1 x2
 		// ditto vt. growth
@@ -171,8 +179,8 @@ public final class Rect implements Geometric {
 	public Point nearestPointTo(Point pt, boolean edgeOnly) {
 		if (pt == null) throw new IllegalArgumentException("null pt");
 
-		float x = pt.x;
-		float y = pt.y;
+		double x = pt.x;
+		double y = pt.y;
 		//  0 | 1 | 2
 		// ---+---+---
 		//  3 | 4 | 5
@@ -195,10 +203,10 @@ public final class Rect implements Geometric {
 		case 8 : return new Point(maxX, maxY);
 		default:
 			if (edgeOnly) {
-				float dxMin = x - minX;
-				float dxMax = maxX - x;
-				float dyMin = y - minY;
-				float dyMax = maxY - y;
+				double dxMin = x - minX;
+				double dxMax = maxX - x;
+				double dyMin = y - minY;
+				double dyMax = maxY - y;
 				if (dxMin == 0 || dxMax == 0 || dyMin == 0 || dyMax==0) return pt;
 				if (dxMin <= dxMax) {
 					if (dyMin <= dyMax) {
@@ -234,11 +242,11 @@ public final class Rect implements Geometric {
 		return new FloatRange(minY, maxY);
 	}
 
-	public float getDiagonalLength() {
+	public double getDiagonalLength() {
 		return Norm.L2.magnitude(maxX - minX, maxY - minY);
 	}
 
-	public float getPerimeterLength() {
+	public double getPerimeterLength() {
 		return 2 * (maxX - minX + maxY - minY);
 	}
 	// geometric methods
@@ -263,19 +271,18 @@ public final class Rect implements Geometric {
 
 	@Override
 	public int hashCode() {
-		int h = Float.floatToIntBits(minX);
-		h = h * 31 + Float.floatToIntBits(maxX);
-		h = h * 31 + Float.floatToIntBits(minY);
-		h = h * 31 + Float.floatToIntBits(maxY);
+		int h = Double.hashCode(minX);
+		h = h * 31 + Double.hashCode(maxX);
+		h = h * 31 + Double.hashCode(minY);
+		h = h * 31 + Double.hashCode(maxY);
 		return h;
 	}
 
 	@Override
 	public boolean equals(Object obj) {
 		if (obj == this) return true;
-		if (!(obj instanceof Rect)) return false;
-		Rect that = (Rect) obj;
-		if (this.minX != that.minX) return false;
+		if (!(obj instanceof Rect that)) return false;
+        if (this.minX != that.minX) return false;
 		if (this.maxX != that.maxX) return false;
 		if (this.minY != that.minY) return false;
 		if (this.maxY != that.maxY) return false;
