@@ -46,15 +46,15 @@ public final class BezierCurve extends Curve {
 	// fields
 
 	private final int order;
-	private final float[] coords;
+	private final double[] coords;
 	private final List<Point> points;
 
-	BezierCurve(List<Point> points, float[] coords) {
+	BezierCurve(List<Point> points, double[] coords) {
 		int size = points.size();
 		if (size == 0) throw new IllegalStateException();
 		int order = size - 1;
 		if (coords == null) {
-			coords = new float[size * 2];
+			coords = new double[size * 2];
 			for (int i = 0, j = 0; i < size; i++) {
 				Point point = points.get(i);
 				coords[j++] = point.x;
@@ -68,13 +68,13 @@ public final class BezierCurve extends Curve {
 
 	private BezierCurve(Point from, Point to) {
 		points = GeomUtil.asList(from, to);
-		coords = new float[] { from.x, from.y, to.x, to.y };
+		coords = new double[] { from.x, from.y, to.x, to.y };
 		order = 1;
 	}
 
 	private BezierCurve(Point only) {
 		points = Collections.singletonList(only);
-		coords = new float[] { only.x, only.y };
+		coords = new double[] { only.x, only.y };
 		order = 0;
 	}
 
@@ -87,34 +87,35 @@ public final class BezierCurve extends Curve {
 	}
 
 	@Override
-	public Point pointAt(float t) {
+	public Point pointAt(double t) {
 		if (t <= 0) return points.get(0);
 		if (t >= 1) return points.get(order);
-		switch (order) {
-		case 0 : return points.get(0);
-		case 1 : return Point.Util.interpolate(points.get(0), points.get(1), t);
-		default:
-			float[] fs = computeCasteljau(t, coords);
-			return new Point(fs[0], fs[1]);
-		}
+        return switch (order) {
+            case 0 -> points.get(0);
+            case 1 -> Point.Util.interpolate(points.get(0), points.get(1), t);
+            default -> {
+                double[] fs = computeCasteljau(t, coords);
+                yield new Point(fs[0], fs[1]);
+            }
+        };
 	}
 
 	@Override
-	public Vector tangentAt(float t) {
+	public Vector tangentAt(double t) {
 		switch (order) {
 		case 0 : return Vector.ZERO;
 		case 1 : return points.get(0).vectorTo(points.get(1)).normalized();
 		default:
 			t = clamp(t);
-			if (t == 0f) return points.get(1).vectorFrom(points.get(0)).normalized();
-			if (t == 1f) return points.get(order).vectorFrom(points.get(order - 1)).normalized();
-			float[] fs = computeCasteljau(clamp(t), coords);
+			if (t == 0.0) return points.get(1).vectorFrom(points.get(0)).normalized();
+			if (t == 1.0) return points.get(order).vectorFrom(points.get(order - 1)).normalized();
+			double[] fs = computeCasteljau(clamp(t), coords);
 			return new Vector(fs[2] - fs[0], fs[3] - fs[1]).normalized();
 		}
 	}
 
 	@Override
-	public PointPath pointTangentAt(float t) {
+	public PointPath pointTangentAt(double t) {
 		if (order == 0) {
 			//TODO how to handle this properly
 			return PointPath.from(points.get(0), Vector.UNIT_X);
@@ -123,21 +124,21 @@ public final class BezierCurve extends Curve {
 		t = clamp(t);
 		Point start;
 		Vector tangent;
-		if (t == 0f) {
+		if (t == 0.0) {
 			start = points.get(0);
 			tangent = points.get(1).vectorFrom(start).normalized();
-		} else if (t == 1f) {
+		} else if (t == 1.0) {
 			start = points.get(order);
 			tangent = points.get(order - 1).vectorTo(start).normalized();
 		} else {
-			float[] fs = computeCasteljau(t, coords);
+			double[] fs = computeCasteljau(t, coords);
 			start = new Point(fs[0], fs[1]);
 			tangent = new Vector(fs[2] - fs[0], fs[3] - fs[1]).normalized();
 		}
 		return PointPath.from(start, tangent);
 	}
 
-	public float magnitudeAt(float t) {
+	public double magnitudeAt(double t) {
 		Norm norm = Norm.L2;
 		if (t <= 0) return norm.distanceFromOrigin(points.get(0));
 		if (t >= 1) return norm.distanceFromOrigin(points.get(order));
@@ -145,41 +146,41 @@ public final class BezierCurve extends Curve {
 		case 0 : return norm.distanceFromOrigin(points.get(0));
 		case 1 : return norm.distanceFromOrigin(Point.Util.interpolate(points.get(0), points.get(1), t));
 		default:
-			float[] fs = computeCasteljau(t, coords);
+			double[] fs = computeCasteljau(t, coords);
 			return norm.magnitude(fs[0], fs[1]);
 		}
 	}
 
 	@Override
 	// note: splits into smaller beziers
-	public SplitCurvePath splitAt(float t) {
+	public SplitCurvePath splitAt(double t) {
 		// TODO consider a pointpath?
 		boolean closed = getPath().isClosed();
 		if (t <= 0) return new SplitCurvePath(new BezierCurve(GeomUtil.asList(points.get(0)), null).getPath(), getPath(), closed);
 		if (t >= 1) return new SplitCurvePath(getPath(), new BezierCurve(GeomUtil.asList(points.get(order)), null).getPath(), closed);
-		float[] fs = computeCasteljau(t, coords);
+		double[] fs = computeCasteljau(t, coords);
 		return new SplitCurvePath(curveTo(t, fs).getPath(), curveFrom(t, fs).getPath(), closed);
 	}
 
 	@Override
-	protected BezierCurve computeSegment(float minP, float maxP) {
+	protected BezierCurve computeSegment(double minP, double maxP) {
 		BezierCurve tail = fromPoints(computeCasteljau(minP, coords).clone(), true);
-		BezierCurve head = fromPoints(computeCasteljau((maxP - minP)/(1f - minP), tail.coords).clone(), false);
+		BezierCurve head = fromPoints(computeCasteljau((maxP - minP)/(1.0 - minP), tail.coords).clone(), false);
 		return head;
 	}
 
 	@Override
 	protected Curve computeDerivative() {
 		List<Point> pts = new ArrayList<>(order);
-		float[] cs = new float[order * 2];
-		float[] coords = this.coords;
+		double[] cs = new double[order * 2];
+		double[] coords = this.coords;
 		for (int i = 0; i < cs.length;) {
-			float ax = coords[i    ];
-			float ay = coords[i + 1];
-			float bx = coords[i + 2];
-			float by = coords[i + 3];
-			float x = (bx - ax) * order;
-			float y = (by - ay) * order;
+			double ax = coords[i    ];
+			double ay = coords[i + 1];
+			double bx = coords[i + 2];
+			double by = coords[i + 3];
+			double x = (bx - ax) * order;
+			double y = (by - ay) * order;
 			cs[i++] = x;
 			cs[i++] = y;
 			pts.add(new Point(x,y));
@@ -202,11 +203,11 @@ public final class BezierCurve extends Curve {
 		return new BezierPath(this);
 	}
 
-	// returns working floats
-	private static float[] computeCasteljau(float t, float[] coords) {
-		final float s = 1 - t;
+	// returns working doubles
+	private static double[] computeCasteljau(double t, double[] coords) {
+		final double s = 1 - t;
 		int length = coords.length;
-		float[] fs = GeomUtil.workingFloats(length * 2);
+		double[] fs = GeomUtil.workingFloats(length * 2);
 		System.arraycopy(coords, 0, fs, 0, length);
 		int j = length;
 		for (length -= 2; length > 0; length -=2) {
@@ -222,7 +223,7 @@ public final class BezierCurve extends Curve {
 		return fs;
 	}
 
-	private BezierCurve curveFrom(float t, float[] fs) {
+	private BezierCurve curveFrom(double t, double[] fs) {
 		if (t <= 0) return this;
 		if (t >= 1) return new BezierCurve(points.get(order));
 		switch (order) {
@@ -237,7 +238,7 @@ public final class BezierCurve extends Curve {
 		}
 	}
 
-	private BezierCurve curveTo(float t, float[] fs) {
+	private BezierCurve curveTo(double t, double[] fs) {
 		if (t <= 0) return new BezierCurve(points.get(0));
 		if (t >= 1) return this;
 		switch (order) {
@@ -252,7 +253,7 @@ public final class BezierCurve extends Curve {
 		}
 	}
 
-	private BezierCurve fromPoints(float[] fs, boolean from) {
+	private BezierCurve fromPoints(double[] fs, boolean from) {
 		Point[] pts = new Point[order + 1];
 		int offset = from ? 0 : order * 2 + 2;
 		for (int i = 0; i < pts.length; i++) {
@@ -297,7 +298,7 @@ public final class BezierCurve extends Curve {
 				consumer.addPoint(z.points.get(0));
 				return consumer.addPoint(z.points.get(1));
 			default:
-				SplitCurvePath curves = ((BezierCurve)z).splitAt(0.5f);
+				SplitCurvePath curves = ((BezierCurve)z).splitAt(0.5);
 				curves.getFirstPath().simplifyCurve().linearize(consumer);
 				return curves.getLastPath().simplifyCurve().linearize(consumer);
 			}
@@ -310,7 +311,7 @@ public final class BezierCurve extends Curve {
 			case 0 : return new SimplifiedPath(LineSegment.fromPoint(z.points.get(0), Vector.UNIT_X).getPath());
 			case 1 : return new SimplifiedPath(LineSegment.fromPoints(z.points.get(0), z.points.get(1)).getPath());
 			default:
-				SplitCurvePath curves = ((BezierCurve)z).splitAt(0.5f);
+				SplitCurvePath curves = ((BezierCurve)z).splitAt(0.5);
 				Path p1 = curves.getFirstPath().simplifyCurve();
 				Path p2 = curves.getLastPath().simplifyCurve();
 				return new SimplifiedPath(new SplitPath(p1, p2, isClosed()));
@@ -340,11 +341,11 @@ public final class BezierCurve extends Curve {
 		public CurvePath apply(Transform t) {
 			if (t == null) throw new IllegalArgumentException("null t");
 			if (t.isIdentity()) return this;
-			float[] coords = z.coords.clone();
+			double[] coords = z.coords.clone();
 			t.transform(coords);
 			Point[] points = new Point[z.order + 1];
 			for (int i = 0, j = 0; i < points.length; i++, j++) {
-				points[i] = new Point(coords[j], coords[j++]);
+				points[i] = new Point(coords[j], coords[++j]);
 			}
 			return new BezierCurve(GeomUtil.asList(points), coords).getPath();
 		}
@@ -359,9 +360,8 @@ public final class BezierCurve extends Curve {
 		@Override
 		public boolean equals(Object obj) {
 			if (obj == this) return true;
-			if (!(obj instanceof BezierPath)) return false;
-			BezierPath that = (BezierPath) obj;
-			return this.z.points.equals(that.z.points);
+			if (!(obj instanceof BezierPath that)) return false;
+            return this.z.points.equals(that.z.points);
 		}
 
 		@Override
