@@ -5,22 +5,22 @@ public abstract class Norm {
 	public static final Norm L1 = new Norm() {
 
 		@Override
-		public float magnitude(float x, float y) {
+		public double magnitude(double x, double y) {
 			return Math.abs(x) + Math.abs(y);
 		}
 
 		@Override
-		public float powMagnitude(float x, float y) {
+		public double powMagnitude(double x, double y) {
 			return Math.abs(x) + Math.abs(y);
 		}
 
 		@Override
-		public float magnitude(int x, int y) {
+		public double magnitude(int x, int y) {
 			return Math.abs(x) + Math.abs(y);
 		}
 
 		@Override
-		public float powMagnitude(int x, int y) {
+		public double powMagnitude(int x, int y) {
 			return Math.abs(x) + Math.abs(y);
 		}
 
@@ -29,26 +29,26 @@ public abstract class Norm {
 	public static final Norm L2 = new Norm() {
 
 		@Override
-		public float magnitude(int x, int y) {
+		public double magnitude(int x, int y) {
 			if (x == 0) return Math.abs(y);
 			if (y == 0) return Math.abs(x);
-			return (float) Math.sqrt(x * x + y * y);
+			return Math.sqrt(x * x + y * y);
 		}
 
 		@Override
-		public float powMagnitude(int x, int y) {
+		public double powMagnitude(int x, int y) {
 			return x * x + y * y;
 		}
 
 		@Override
-		public float magnitude(float x, float y) {
-			if (x == 0f) return Math.abs(y);
-			if (y == 0f) return Math.abs(x);
-			return (float) Math.sqrt(x * x + y * y);
+		public double magnitude(double x, double y) {
+			if (x == 0.0) return Math.abs(y);
+			if (y == 0.0) return Math.abs(x);
+			return Math.sqrt(x * x + y * y);
 		}
 
 		@Override
-		public float powMagnitude(float x, float y) {
+		public double powMagnitude(double x, double y) {
 			return x * x + y * y;
 		}
 
@@ -58,7 +58,7 @@ public abstract class Norm {
 		}
 
 		@Override
-		public float magnitude(Vector vector) {
+		public double magnitude(Vector vector) {
 			return vector.getMagnitude();
 		}
 
@@ -67,23 +67,23 @@ public abstract class Norm {
 	public static final Norm Linf = new Norm() {
 
 		@Override
-		public float magnitude(float x, float y) {
+		public double magnitude(double x, double y) {
 			return Math.max(x, y);
 		}
 
 		@Override
-		public float powMagnitude(int x, int y) {
-			return x == 0 && y == 0 ? 0 : Float.POSITIVE_INFINITY;
+		public double powMagnitude(int x, int y) {
+			return x == 0 && y == 0 ? 0 : Double.POSITIVE_INFINITY;
 		}
 
 		@Override
-		public float magnitude(int x, int y) {
+		public double magnitude(int x, int y) {
 			return Math.max(x, y);
 		}
 
 		@Override
-		public float powMagnitude(float x, float y) {
-			return x == 0f && y == 0f ? 0 : Float.POSITIVE_INFINITY;
+		public double powMagnitude(double x, double y) {
+			return x == 0.0 && y == 0.0 ? 0 : Double.POSITIVE_INFINITY;
 		}
 
 	};
@@ -95,54 +95,54 @@ public abstract class Norm {
 		if (p == Double.POSITIVE_INFINITY) return Linf;
 		return new Norm() {
 			@Override
-			public float magnitude(float x, float y) {
-				if (x == 0f) return Math.abs(y);
-				if (y == 0f) return Math.abs(x);
-				return (float) Math.pow( Math.pow(Math.abs(x), p) + Math.pow(Math.abs(y), p), 1/p );
+			public double magnitude(double x, double y) {
+				if (x == 0.0) return Math.abs(y);
+				if (y == 0.0) return Math.abs(x);
+				return Math.pow( Math.pow(Math.abs(x), p) + Math.pow(Math.abs(y), p), 1/p );
 			}
 
 			@Override
-			public float powMagnitude(int x, int y) {
-				return (float) Math.pow(Math.abs(x), p) + (float) Math.pow(Math.abs(y), p);
+			public double powMagnitude(int x, int y) {
+				return Math.pow(Math.abs(x), p) + Math.pow(Math.abs(y), p);
 			}
 
 			@Override
-			public float magnitude(int x, int y) {
-				return magnitude(x, y);
+			public double magnitude(int x, int y) {
+				return magnitude((double) x, (double) y);
 			}
 
 			@Override
-			public float powMagnitude(float x, float y) {
-				return (float) Math.pow(Math.abs(x), p) + (float) Math.pow(Math.abs(y), p);
+			public double powMagnitude(double x, double y) {
+				return Math.pow(Math.abs(x), p) + Math.pow(Math.abs(y), p);
 			}
 		};
 	}
 
-	public float distanceFromOrigin(Point point) {
+	public double distanceFromOrigin(Point point) {
 		return magnitude(point.x, point.y);
 	}
 
-	public float powDistanceFromOrigin(Point point) {
+	public double powDistanceFromOrigin(Point point) {
 		return powMagnitude(point.x, point.y);
 	}
 
-	public float lengthOf(LinearPath path) {
+	public double lengthOf(LinearPath path) {
 		return distanceBetween(path.getStart(), path.getFinish());
 	}
 
-	public float powLengthOf(LinearPath path) {
+	public double powLengthOf(LinearPath path) {
 		return powDistanceBetween(path.getStart(), path.getFinish());
 	}
 
-	public float distanceBetween(Point pt1, Point pt2) {
+	public double distanceBetween(Point pt1, Point pt2) {
 		return magnitude(pt1.x - pt2.x, pt1.y - pt2.y);
 	}
 
-	public float powDistanceBetween(Point pt1, Point pt2) {
+	public double powDistanceBetween(Point pt1, Point pt2) {
 		return powMagnitude(pt1.x - pt2.x, pt1.y - pt2.y);
 	}
 
-	public float magnitude(Vector vector) {
+	public double magnitude(Vector vector) {
 		return magnitude(vector.x, vector.y);
 	}
 
@@ -157,12 +157,12 @@ public abstract class Norm {
 		return LineSegment.fromVector(start, v);
 	}
 
-	public abstract float magnitude(int x, int y);
+	public abstract double magnitude(int x, int y);
 
-	public abstract float powMagnitude(int x, int y);
+	public abstract double powMagnitude(int x, int y);
 
-	public abstract float magnitude(float x, float y);
+	public abstract double magnitude(double x, double y);
 
-	public abstract float powMagnitude(float x, float y);
+	public abstract double powMagnitude(double x, double y);
 
 }
