@@ -11,36 +11,36 @@ import com.tomgibara.geom.transform.Transform;
 
 public final class Ellipse implements Geometric {
 
-	private static final Rect BASIC_BOUND = Rect.atPoints(-1f, -1f, 1f, 1f);
+	private static final Rect BASIC_BOUND = Rect.atPoints(-1.0, -1.0, 1.0, 1.0);
 
 	public static Ellipse fromTransform(Transform t) {
 		if (t == null) throw new IllegalArgumentException("null t");
 		return new Ellipse(t);
 	}
 
-	public static Ellipse fromRadius(float radius) {
-		if (radius == 0f) throw new IllegalArgumentException("zero radius");
+	public static Ellipse fromRadius(double radius) {
+		if (radius == 0.0) throw new IllegalArgumentException("zero radius");
 		return new Ellipse( Transform.scale(radius) );
 	}
 
-	public static Ellipse fromRadius(Point center, float radius) {
+	public static Ellipse fromRadius(Point center, double radius) {
 		if (center == null) throw new IllegalArgumentException("null center");
-		if (radius == 0f) throw new IllegalArgumentException("zero radius");
+		if (radius == 0.0) throw new IllegalArgumentException("zero radius");
 		Transform t = Transform.scale(radius);
 		if (!center.isOrigin()) t = t.apply(center.vectorFromOrigin().asTranslation());
 		return new Ellipse(t);
 	}
 
-	public static Ellipse fromRadii(float xRadius, float yRadius) {
-		if (xRadius == 0f) throw new IllegalArgumentException("zero xRadius");
-		if (yRadius == 0f) throw new IllegalArgumentException("zero yRadius");
+	public static Ellipse fromRadii(double xRadius, double yRadius) {
+		if (xRadius == 0.0) throw new IllegalArgumentException("zero xRadius");
+		if (yRadius == 0.0) throw new IllegalArgumentException("zero yRadius");
 		return new Ellipse( Transform.scale(xRadius, yRadius) );
 	}
 
-	public static Ellipse fromRadii(Point center, float xRadius, float yRadius) {
+	public static Ellipse fromRadii(Point center, double xRadius, double yRadius) {
 		if (center == null) throw new IllegalArgumentException("null center");
-		if (xRadius == 0f) throw new IllegalArgumentException("zero xRadius");
-		if (yRadius == 0f) throw new IllegalArgumentException("zero yRadius");
+		if (xRadius == 0.0) throw new IllegalArgumentException("zero xRadius");
+		if (yRadius == 0.0) throw new IllegalArgumentException("zero yRadius");
 		Transform t = Transform.scale(xRadius, yRadius);
 		if (!center.isOrigin()) t = t.apply(center.vectorFromOrigin().asTranslation());
 		return new Ellipse(t);
@@ -48,7 +48,7 @@ public final class Ellipse implements Geometric {
 
 	public static Ellipse fromRect(Rect rect) {
 		if (rect == null) throw new IllegalArgumentException("null rect");
-		return fromRadii(rect.getCenter(), rect.getWidth() * 0.5f, rect.getHeight() * 0.5f);
+		return fromRadii(rect.getCenter(), rect.getWidth() * 0.5, rect.getHeight() * 0.5);
 	}
 
 	private final Transform transform;
@@ -58,9 +58,9 @@ public final class Ellipse implements Geometric {
 	private Vector semiMinorAxis = null;
 	private Vector semiMajorAxis = null;
 	//private boolean firstIsMajor;
-	private float orientation;
-	private float maxR = Float.NaN;
-	private float minR = Float.NaN;
+	private double orientation;
+	private double maxR = Double.NaN;
+	private double minR = Double.NaN;
 	private Rect bounds = null;
 
 	private Ellipse(Transform transform) {
@@ -81,13 +81,13 @@ public final class Ellipse implements Geometric {
 		return center == null ? center = transform.transformOrigin() : center;
 	}
 
-	public float getMajorRadius() {
-		if (Float.isNaN(maxR)) computeRadii();
+	public double getMajorRadius() {
+		if (Double.isNaN(maxR)) computeRadii();
 		return maxR;
 	}
 
-	public float getMinorRadius() {
-		if (Float.isNaN(minR)) computeRadii();
+	public double getMinorRadius() {
+		if (Double.isNaN(minR)) computeRadii();
 		return minR;
 	}
 
@@ -101,12 +101,12 @@ public final class Ellipse implements Geometric {
 		return semiMinorAxis;
 	}
 
-	public EllipticalArc arc(float startAngle, float finishAngle) {
+	public EllipticalArc arc(double startAngle, double finishAngle) {
 		return new EllipticalArc(this, startAngle / Angles.TWO_PI, finishAngle / Angles.TWO_PI);
 	}
 
 	public EllipticalArc completeArc() {
-		return new EllipticalArc(this, 0f, 1f);
+		return new EllipticalArc(this, 0.0, 1.0);
 	}
 
 	@Override
@@ -114,10 +114,10 @@ public final class Ellipse implements Geometric {
 		return new Ellipse(transform.apply(t));
 	}
 
-	public Point getPoint(float scaledAngle) {
-		float angle = scaledAngle * TWO_PI;
-		float c = (float) Math.cos(angle);
-		float s = (float) Math.sin(angle);
+	public Point getPoint(double scaledAngle) {
+		double angle = scaledAngle * TWO_PI;
+		double c = Math.cos(angle);
+		double s = Math.sin(angle);
 		return transform.transform(c, s);
 	}
 
@@ -126,15 +126,15 @@ public final class Ellipse implements Geometric {
 		return bounds == null ? bounds = computeBounds() : bounds;
 	}
 
-	public Rect getArcBounds(float startAngleScaled, float finishAngleScaled) {
+	public Rect getArcBounds(double startAngleScaled, double finishAngleScaled) {
 		Rect b1 = getBounds();
 		//TODO
-//		if (Math.abs(finishAngleScaled - startAngleScaled) >= 0.5f) return b1;
+//		if (Math.abs(finishAngleScaled - startAngleScaled) >= 0.5) return b1;
 //		// this is fairly efficient for small arcs
 //		Point pt1 = getPoint(startAngleScaled);
 //		Point pt2 = getPoint(finishAngleScaled);
 //		Point c = Point.Util.midpoint(pt1, pt2);
-//		float r = Norm.L2.distanceBetween(c, pt1);
+//		double r = Norm.L2.distanceBetween(c, pt1);
 //		Rect b2 = Rect.atPoints(c.x - r, c.y - r, c.x + r, c.y + r);
 //		return Rect.intersectionRect(b1, b2);
 		return b1;
@@ -169,8 +169,8 @@ public final class Ellipse implements Geometric {
 	private void computeRadii() {
 //		Transform basis = getEigenBasis();
 //		//TODO is there a way to avoid these sqrts?
-//		float r1 = basis.getColumn(0).getMagnitude();
-//		float r2 = basis.getColumn(1).getMagnitude();
+//		double r1 = basis.getColumn(0).getMagnitude();
+//		double r2 = basis.getColumn(1).getMagnitude();
 //		if (basis.isFirstColumnMajor()) {
 //			maxR = r1;
 //			minR = r2;
@@ -179,8 +179,8 @@ public final class Ellipse implements Geometric {
 //			minR = r1;
 //		}
 		Vector es = transform.getEigenValues();
-		float e1 = Math.abs(es.x);
-		float e2 = Math.abs(es.y);
+		double e1 = Math.abs(es.x);
+		double e2 = Math.abs(es.y);
 		if (e2 > e1) {
 			maxR = e2;
 			minR = e1;
@@ -191,26 +191,26 @@ public final class Ellipse implements Geometric {
 	}
 
 	private void computeSemiAxes() {
-		float m00 = transform.m00;
-		float m01 = transform.m01;
-		float m10 = transform.m10;
-		float m11 = transform.m11;
+		double m00 = transform.m00;
+		double m01 = transform.m01;
+		double m10 = transform.m10;
+		double m11 = transform.m11;
 
-		float a = -(m00 * m01 + m10 * m11);
-		float b = (m11 + m00) * (m11 - m00);
-		float c = (m00 * m10 + m01 * m11);
+		double a = -(m00 * m01 + m10 * m11);
+		double b = (m11 + m00) * (m11 - m00);
+		double c = (m00 * m10 + m01 * m11);
 
-		float d = (float) Math.sqrt(b*b - 4 * a * c);
-//		float t1 = (-b + d) / (a * 2f);
-//		float t2 = (-b - d) / (a * 2f);
+		double d = Math.sqrt(b*b - 4 * a * c);
+//		double t1 = (-b + d) / (a * 2.0);
+//		double t2 = (-b - d) / (a * 2.0);
 
-		//float angle = (float) Math.atan2(-b + d, a * 2f);
-		semiMajorAxis = new Vector(a * 2f, -b + d).normalized().scaled(50f);
-		semiMinorAxis = new Vector(a * 2f, -b - d).normalized().scaled(50f);
+		//double angle = Math.atan2(-b + d, a * 2.0);
+		semiMajorAxis = new Vector(a * 2.0, -b + d).normalized().scaled(50.0);
+		semiMinorAxis = new Vector(a * 2.0, -b - d).normalized().scaled(50.0);
 
 		Vector vx = transform.transform(Vector.UNIT_X);
 		Vector vz = transform.transform(Vector.ZERO);
-		semiMajorAxis = vx.subtract(vz).normalized().scaled(50f);
+		semiMajorAxis = vx.subtract(vz).normalized().scaled(50.0);
 		semiMinorAxis = Vector.ZERO;
 
 	}
