@@ -10,15 +10,15 @@ import com.tomgibara.geom.path.Path;
 
 public class PatternDash implements Dash {
 
-	public static PatternDash single(float dashLength, float gapLength) {
-		return new PatternDash(dashLength + gapLength, new FloatRange(0f, dashLength));
+	public static PatternDash single(double dashLength, double gapLength) {
+		return new PatternDash(dashLength + gapLength, new FloatRange(0.0, dashLength));
 	}
 
 	//TODO could allow empty ranges?
-	public static PatternDash pattern(float length, FloatRange... ranges) {
+	public static PatternDash pattern(double length, FloatRange... ranges) {
 		if (ranges == null) throw new IllegalArgumentException("null ranges");
 		if (ranges.length == 0) throw new IllegalArgumentException("no ranges");
-		float min = 0f;
+		double min = 0.0;
 		for (int i = 0; i < ranges.length; i++) {
 			FloatRange range = ranges[i];
 			if (!range.isClosed() || range.isEmpty() || range.min < min) throw new IllegalArgumentException("invalid range");
@@ -28,15 +28,15 @@ public class PatternDash implements Dash {
 		return new PatternDash(length, ranges.clone());
 	}
 
-	private final float length;
+	private final double length;
 	private final FloatRange[] ranges;
 
-	private PatternDash(float length, FloatRange... ranges) {
+	private PatternDash(double length, FloatRange... ranges) {
 		this.length = length;
 		this.ranges = ranges;
 	}
 
-	public float getLength() {
+	public double getLength() {
 		return length;
 	}
 
@@ -48,7 +48,7 @@ public class PatternDash implements Dash {
 	public Iterator<FloatRange> getPattern(Path path) {
 		return new Iterator<FloatRange>() {
 
-			float offset = 0f;
+			double offset = 0.0;
 			int index = 0;
 
 			@Override
